Separated convertRelative
